correct 0-linear

+++ 0x1E-search_algorithms/0-linear.c
@@ -13,14 +13,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i = 0;
+	size_t i = 0;
 
 	if (!array)
 		return (-1);
 
 	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%d] = [%d]\n", i, array[i]);
+		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 
 		if (array[i] == value)
 			return (i);
